rename SVMInput -> ClassificationInput
